setup get planned meals and sync with firestore
