+++ app/src/main/java/com/example/mealwise/di/Injection.java
@@ -6,8 +6,8 @@ import com.example.mealwise.data.auth.datasource.AuthLocalDataSource;
 import com.example.mealwise.data.auth.datasource.AuthLocalDataSourceImpl;
 import com.example.mealwise.data.auth.datasource.AuthRemoteDataSource;
 import com.example.mealwise.data.auth.datasource.AuthRemoteDataSourceImp;
-import com.example.mealwise.data.auth.datasource.helpers.FirestoreHelper;
-import com.example.mealwise.data.auth.datasource.helpers.SharedPrefHelper;
+import com.example.mealwise.utils.helpers.FirestoreHelper;
+import com.example.mealwise.utils.helpers.SharedPrefHelper;
 import com.example.mealwise.data.auth.repository.AuthRepository;
 import com.example.mealwise.data.auth.repository.AuthRepositoryImpl;
 import com.example.mealwise.data.meals.datasource.local.MealsLocalDataSource;
