+++ app/src/main/java/com/example/mealwise/presentation/auth/signIn/view/SignInFragment.java
@@ -13,9 +13,7 @@ import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 
 import com.example.mealwise.R;
-import com.example.mealwise.data.auth.datasource.AuthRemoteDataSourceImp;
-import com.example.mealwise.data.auth.datasource.helpers.SharedPrefHelper;
-import com.example.mealwise.data.auth.repository.AuthRepositoryImpl;
+import com.example.mealwise.utils.helpers.SharedPrefHelper;
 import com.example.mealwise.di.Injection;
 import com.example.mealwise.presentation.auth.base.BaseAuthFragment;
 import com.example.mealwise.presentation.auth.signIn.presenter.SignInPresenter;
