+++ app/src/main/java/com/example/mealwise/presentation/splash/view/SplashFragment.java
@@ -14,9 +14,6 @@ import android.view.ViewGroup;
 import android.view.WindowManager;
 
 import com.example.mealwise.R;
-import com.example.mealwise.data.auth.datasource.AuthRemoteDataSourceImp;
-import com.example.mealwise.data.auth.datasource.helpers.SharedPrefHelper;
-import com.example.mealwise.data.auth.repository.AuthRepositoryImpl;
 import com.example.mealwise.di.Injection;
 import com.example.mealwise.presentation.splash.presenter.SplashPresenter;
 import com.example.mealwise.presentation.splash.presenter.SplashPresenterImpl;
