+++ app/src/main/java/com/example/mealwise/utils/helpers/SharedPrefHelper.java
@@ -1,4 +1,4 @@
-package com.example.mealwise.data.auth.datasource.helpers;
+package com.example.mealwise.utils.helpers;
 
 import android.content.Context;
 import android.content.SharedPreferences;
